Update main.cpp and helper

+++ helper/Helper.hpp
@@ -7,7 +7,7 @@ void InputArrayFromTxtFile(int *&a, int &n, const char *fileName);
 void WriteArrayToTxtFile(int *a, int n, const char *fileName);
 void PrintArray(int a[], int n);
 void PrintResult(const char *outPara, int runningTime, ull countCompare);
-int MeasureRunningTime(std::function<void(int *, int)> algoFunc, int a[], int size);
+ull MeasureRunningTime(std::function<void(int *, int)> algoFunc, int a[], int size);
 void PrintOrderName(int orderIdx);
 void WriteExperimentalResultToCsvFile(std::string algoName, int orderIdx, int inputSize , int runningTime, ull countCompare);
